Implementação da tela de pagamento + criação da procedure de cadastro de pedido

+++ VLDocesWeb/Models/CartItemModel.cs
@@ -3,7 +3,7 @@ namespace VLDocesWeb.Models;
 public class CartItem
 {
     public Product Produto { set; get; }
-    public float PrecoVendido { set; get; }
+    public decimal PrecoVendido { set; get; }
     public int Quantidade { set; get; }
     public int PedidoId { set; get; }
 }
